[#763] Finishing touches for CompactBlockProcessor refactor
Closes #763
